backend/requirements.txt
Django===5.0.6
djangorestframework
gunicorn
